+++ cdn/hot-connect.es.js
@@ -9,42 +9,42 @@ class k {
     typeof window > "u" || localStorage.removeItem(e);
   }
 }
-const m = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz";
-function S(r) {
+const f = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz";
+function I(r) {
   if (r.length === 0) return "";
   let e = 0, t = 0;
   for (; t < r.length && r[t] === 0; )
     e++, t++;
   let n = [0];
   for (; t < r.length; t++) {
-    let s = r[t];
+    let o = r[t];
     for (let a = 0; a < n.length; ++a)
-      s += n[a] << 8, n[a] = s % 58, s = s / 58 | 0;
-    for (; s > 0; )
-      n.push(s % 58), s = s / 58 | 0;
+      o += n[a] << 8, n[a] = o % 58, o = o / 58 | 0;
+    for (; o > 0; )
+      n.push(o % 58), o = o / 58 | 0;
   }
   for (; n.length > 0 && n[n.length - 1] === 0; ) n.pop();
-  let o = "";
-  for (let s = 0; s < e; s++)
-    o += m[0];
-  for (let s = n.length - 1; s >= 0; --s)
-    o += m[n[s]];
-  return o;
+  let s = "";
+  for (let o = 0; o < e; o++)
+    s += f[0];
+  for (let o = n.length - 1; o >= 0; --o)
+    s += f[n[o]];
+  return s;
 }
-const C = (r) => {
+const S = (r) => {
   try {
     return JSON.parse(new TextDecoder().decode(r));
   } catch {
     return r;
   }
-}, d = (r) => r.map((e) => {
+}, h = (r) => r.map((e) => {
   if ("type" in e) return e;
   if (e.functionCall)
     return {
       type: "FunctionCall",
       params: {
         methodName: e.functionCall.methodName,
-        args: C(e.functionCall.args),
+        args: S(e.functionCall.args),
         gas: e.functionCall.gas.toString(),
         deposit: e.functionCall.deposit.toString()
       }
@@ -63,7 +63,7 @@ const C = (r) => {
     return {
       type: "UseGlobalContract",
       params: {
-        contractIdentifier: e.useGlobalContract.contractIdentifier.AccountId ? { accountId: e.useGlobalContract.contractIdentifier.AccountId } : { codeHash: S(e.useGlobalContract.contractIdentifier.CodeHash) }
+        contractIdentifier: e.useGlobalContract.contractIdentifier.AccountId ? { accountId: e.useGlobalContract.contractIdentifier.AccountId } : { codeHash: I(e.useGlobalContract.contractIdentifier.CodeHash) }
       }
     };
   if (e.deployContract)
@@ -114,15 +114,15 @@ const C = (r) => {
   const e = Math.random() * 16 | 0;
   return (r === "x" ? e : e & 3 | 8).toString(16);
 });
-class I {
+class A {
   constructor(e, t) {
     this.connector = e, this.manifest = t;
   }
   callParentFrame(e, t) {
     const n = g();
-    return window.parent.postMessage({ type: "near-wallet-injected-request", id: n, method: e, params: t }, "*"), new Promise((o, s) => {
+    return window.parent.postMessage({ type: "near-wallet-injected-request", id: n, method: e, params: t }, "*"), new Promise((s, o) => {
       const a = (i) => {
-        i.data.type === "near-wallet-injected-response" && i.data.id === n && (window.removeEventListener("message", a), i.data.success ? o(i.data.result) : s(i.data.error));
+        i.data.type === "near-wallet-injected-response" && i.data.id === n && (window.removeEventListener("message", a), i.data.success ? s(i.data.result) : o(i.data.error));
       };
       window.addEventListener("message", a);
     });
@@ -135,6 +135,15 @@ class I {
     });
     return Array.isArray(t) ? t : [t];
   }
+  async signInAndSignMessage(e) {
+    const t = await this.callParentFrame("near:signInAndSignMessage", {
+      network: e?.network || this.connector.network,
+      contractId: e?.contractId,
+      methodNames: e?.methodNames,
+      messageParams: e.messageParams
+    });
+    return Array.isArray(t) ? t : [t];
+  }
   async signOut(e) {
     const t = { ...e, network: e?.network || this.connector.network };
     await this.callParentFrame("near:signOut", t);
@@ -144,13 +153,13 @@ class I {
     return this.callParentFrame("near:getAccounts", t);
   }
   async signAndSendTransaction(e) {
-    const t = d(e.actions), n = { ...e, actions: t, network: e.network || this.connector.network };
+    const t = h(e.actions), n = { ...e, actions: t, network: e.network || this.connector.network };
     return this.callParentFrame("near:signAndSendTransaction", n);
   }
   async signAndSendTransactions(e) {
     const t = { ...e, network: e.network || this.connector.network };
     return t.transactions = t.transactions.map((n) => ({
-      actions: d(n.actions),
+      actions: h(n.actions),
       receiverId: n.receiverId
     })), this.callParentFrame("near:signAndSendTransactions", t);
   }
@@ -163,7 +172,7 @@ class I {
       ...e,
       delegateActions: e.delegateActions.map((n) => ({
         ...n,
-        actions: d(n.actions)
+        actions: h(n.actions)
       })),
       network: e.network || this.connector.network
     };
@@ -214,8 +223,8 @@ class y {
    * @param callback Function to be called when event is emitted
    */
   once(e, t) {
-    const n = (o) => {
-      t(o), this.off(e, n);
+    const n = (s) => {
+      t(s), this.off(e, n);
     };
     this.on(e, n);
   }
@@ -228,21 +237,21 @@ class y {
     e ? delete this.events[e] : this.events = {};
   }
 }
-function A(r) {
+function C(r) {
   return r.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;").replace(/'/g, "&#039;");
 }
-const f = Symbol("htmlTag");
-function h(r, ...e) {
+const m = Symbol("htmlTag");
+function d(r, ...e) {
   let t = r[0];
   for (let n = 0; n < e.length; n++) {
-    for (const o of Array.isArray(e[n]) ? e[n] : [e[n]]) {
-      const s = o?.[f] ? o[f] : A(String(o ?? ""));
-      t += s;
+    for (const s of Array.isArray(e[n]) ? e[n] : [e[n]]) {
+      const o = s?.[m] ? s[m] : C(String(s ?? ""));
+      t += o;
     }
     t += r[n + 1];
   }
   return Object.freeze({
-    [f]: t,
+    [m]: t,
     get html() {
       return t;
     }
@@ -521,12 +530,12 @@ class b {
   root = document.createElement("div");
   state = {};
   get dom() {
-    return h``;
+    return d``;
   }
   disposables = [];
   addListener(e, t, n) {
-    const o = typeof e == "string" ? this.root.querySelector(e) : e;
-    o && (o.addEventListener(t, n), this.disposables.push(() => o.removeEventListener(t, n)));
+    const s = typeof e == "string" ? this.root.querySelector(e) : e;
+    s && (s.addEventListener(t, n), this.disposables.push(() => s.removeEventListener(t, n)));
   }
   handlers() {
     this.disposables.forEach((n) => n()), this.disposables = [];
@@ -573,30 +582,25 @@ class E extends b {
   }
   get footer() {
     if (!this.delegate.footer) return "";
-    const { icon: e, heading: t, link: n, linkText: o } = this.delegate.footer;
-    return h`
+    const { icon: e, heading: t } = this.delegate.footer;
+    return d`
       <div class="footer">
-        <img src="${e}" alt="${t}" />
+        ${e ? d`<img src="${e}" alt="${t}" />` : ""}
         <p>${t}</p>
-        <a class="get-wallet-link" href="${n}" target="_blank">${o}</a>
       </div>
     `;
   }
   get dom() {
-    return h` <div class="modal-container">
+    return d`<div class="modal-container">
       <div class="modal-content">
         <div class="modal-body" style="padding: 0; overflow: auto;"></div>
-        <div class="footer">
-          <img src="https://tgapp.herewallet.app/images/hot/hot-icon.png" alt="NEAR Connector" />
-          <p>NEAR Connector</p>
-          <p class="get-wallet-link">Don't have a wallet?</p>
-        </div>
+        ${this.footer}
       </div>
     </div>`;
   }
 }
 async function P(r) {
-  const e = await r.executor.getAllStorage(), t = r.executor.connector.providers, n = r.executor.manifest, o = r.id, s = r.code.replaceAll(".localStorage", ".sandboxedLocalStorage").replaceAll("window.top", "window.selector").replaceAll("window.open", "window.selector.open");
+  const e = await r.executor.getAllStorage(), t = r.executor.connector.providers, n = r.executor.manifest, s = r.id, o = r.code.replaceAll(".localStorage", ".sandboxedLocalStorage").replaceAll("window.top", "window.selector").replaceAll("window.open", "window.selector.open");
   return (
     /* html */
     `
@@ -740,7 +744,7 @@ async function P(r) {
           this.windowIdPromise = window.selector.call("open", { url, features });
 
           window.addEventListener("message", async (event) => {            
-            if (event.data.origin !== "${o}") return;
+            if (event.data.origin !== "${s}") return;
             if (!event.data.method?.startsWith("proxy-window:")) return;
             const method = event.data.method.replace("proxy-window:", "");
             if (method === "closed" && event.data.windowId === await this.id()) this.closed = true;
@@ -805,17 +809,17 @@ async function P(r) {
         },
       
         async ready(wallet) {
-          window.parent.postMessage({ method: "wallet-ready", origin: "${o}" }, "*");
+          window.parent.postMessage({ method: "wallet-ready", origin: "${s}" }, "*");
           window.selector.wallet = wallet;
         },
 
         async call(method, params) {
           const id = window.selector.uuid();
-          window.parent.postMessage({ method, params, id, origin: "${o}" }, "*");
+          window.parent.postMessage({ method, params, id, origin: "${s}" }, "*");
 
           return new Promise((resolve, reject) => {
             const handler = (event) => {
-              if (event.data.id !== id || event.data.origin !== "${o}") return;
+              if (event.data.id !== id || event.data.origin !== "${s}") return;
               window.removeEventListener("message", handler);
 
               if (event.data.status === "failed") reject(event.data.result);
@@ -829,7 +833,7 @@ async function P(r) {
         panelClosed(windowId) {
           window.parent.postMessage({ 
             method: "panel.closed", 
-            origin: "${o}", 
+            origin: "${s}", 
             result: { windowId } 
           }, "*");
         },
@@ -882,12 +886,12 @@ async function P(r) {
       };
 
       window.addEventListener("message", async (event) => {
-        if (event.data.origin !== "${o}") return;
+        if (event.data.origin !== "${s}") return;
         if (!event.data.method?.startsWith("wallet:")) return;
       
         const wallet = window.selector.wallet;
         const method = event.data.method.replace("wallet:", "");
-        const payload = { id: event.data.id, origin: "${o}", method };
+        const payload = { id: event.data.id, origin: "${s}", method };
       
         if (wallet == null || typeof wallet[method] !== "function") {
           const data = { ...payload, status: "failed", result: "Method not found" };
@@ -905,7 +909,7 @@ async function P(r) {
       });
       <\/script>
 
-      <script type="module">${s}<\/script>
+      <script type="module">${o}<\/script>
     </body>
   </html>
     `
@@ -913,12 +917,12 @@ async function P(r) {
 }
 class M {
   constructor(e, t, n) {
-    this.executor = e, this.origin = g(), this.handler = (s) => {
-      s.data.origin === this.origin && (s.data.method === "wallet-ready" && this.readyPromiseResolve(), n(this, s));
+    this.executor = e, this.origin = g(), this.handler = (o) => {
+      o.data.origin === this.origin && (o.data.method === "wallet-ready" && this.readyPromiseResolve(), n(this, o));
     }, window.addEventListener("message", this.handler);
-    const o = [];
-    this.executor.checkPermissions("usb") && o.push("usb *;"), this.executor.checkPermissions("hid") && o.push("hid *;"), this.executor.checkPermissions("clipboardRead") && o.push("clipboard-read;"), this.executor.checkPermissions("clipboardWrite") && o.push("clipboard-write;"), this.iframe.allow = o.join(" "), this.iframe.setAttribute("sandbox", "allow-scripts"), P({ id: this.origin, executor: this.executor, code: t }).then((s) => {
-      this.executor.connector.logger?.log("Iframe code injected"), this.iframe.srcdoc = s;
+    const s = [];
+    this.executor.checkPermissions("usb") && s.push("usb *;"), this.executor.checkPermissions("hid") && s.push("hid *;"), this.executor.checkPermissions("clipboardRead") && s.push("clipboard-read;"), this.executor.checkPermissions("clipboardWrite") && s.push("clipboard-write;"), this.iframe.allow = s.join(" "), this.iframe.setAttribute("sandbox", "allow-scripts"), P({ id: this.origin, executor: this.executor, code: t }).then((o) => {
+      this.executor.connector.logger?.log("Iframe code injected"), this.iframe.srcdoc = o;
     }), this.popup = new E({
       footer: this.executor.connector.footerBranding,
       iframe: this.iframe,
@@ -970,8 +974,8 @@ class L {
       return !n || !t?.entity ? !1 : n.includes(t.entity);
     }
     if (e === "allowsOpen") {
-      const n = u(t?.url || ""), o = this.manifest.permissions.allowsOpen;
-      return !n || !o || !Array.isArray(o) || o.length === 0 ? !1 : o.some((a) => {
+      const n = u(t?.url || ""), s = this.manifest.permissions.allowsOpen;
+      return !n || !s || !Array.isArray(s) || s.length === 0 ? !1 : s.some((a) => {
         const i = u(a);
         return !(!i || n.protocol !== i.protocol || i.hostname && n.hostname !== i.hostname || i.pathname && i.pathname !== "/" && n.pathname !== i.pathname);
       });
@@ -983,10 +987,10 @@ class L {
       throw e.postMessage({ ...n.data, status: "failed", result: "Permission denied" }), new Error("Permission denied");
   }
   _onMessage = async (e, t) => {
-    const n = (s) => {
-      e.postMessage({ ...t.data, status: "success", result: s });
-    }, o = (s) => {
-      e.postMessage({ ...t.data, status: "failed", result: s });
+    const n = (o) => {
+      e.postMessage({ ...t.data, status: "success", result: o });
+    }, s = (o) => {
+      e.postMessage({ ...t.data, status: "failed", result: o });
     };
     if (t.data.method === "ui.showIframe") {
       e.show(), n(null);
@@ -1002,14 +1006,14 @@ class L {
     }
     if (t.data.method === "storage.get") {
       this.assertPermissions(e, "storage", t);
-      const s = localStorage.getItem(`${this.storageSpace}:${t.data.params.key}`);
-      n(s);
+      const o = localStorage.getItem(`${this.storageSpace}:${t.data.params.key}`);
+      n(o);
       return;
     }
     if (t.data.method === "storage.keys") {
       this.assertPermissions(e, "storage", t);
-      const s = Object.keys(localStorage).filter((a) => a.startsWith(`${this.storageSpace}:`));
-      n(s);
+      const o = Object.keys(localStorage).filter((a) => a.startsWith(`${this.storageSpace}:`));
+      n(o);
       return;
     }
     if (t.data.method === "storage.remove") {
@@ -1017,18 +1021,18 @@ class L {
       return;
     }
     if (t.data.method === "panel.focus") {
-      const s = this.activePanels[t.data.params.windowId];
-      s && s.focus(), n(null);
+      const o = this.activePanels[t.data.params.windowId];
+      o && o.focus(), n(null);
       return;
     }
     if (t.data.method === "panel.postMessage") {
-      const s = this.activePanels[t.data.params.windowId];
-      s && s.postMessage(t.data.params.data, "*"), n(null);
+      const o = this.activePanels[t.data.params.windowId];
+      o && o.postMessage(t.data.params.data, "*"), n(null);
       return;
     }
     if (t.data.method === "panel.close") {
-      const s = this.activePanels[t.data.params.windowId];
-      s && s.close(), delete this.activePanels[t.data.params.windowId], n(null);
+      const o = this.activePanels[t.data.params.windowId];
+      o && o.close(), delete this.activePanels[t.data.params.windowId], n(null);
       return;
     }
     if (t.data.method === "walletConnect.connect") {
@@ -1037,16 +1041,16 @@ class L {
         if (!this.connector.walletConnect) throw new Error("WalletConnect is not configured");
         const a = await (await this.connector.walletConnect).connect(t.data.params);
         a.approval(), n({ uri: a.uri });
-      } catch (s) {
-        o(s);
+      } catch (o) {
+        s(o);
       }
       return;
     }
     if (t.data.method === "walletConnect.getProjectId") {
       if (!this.connector.walletConnect) throw new Error("WalletConnect is not configured");
       this.assertPermissions(e, "walletConnect", t);
-      const s = await this.connector.walletConnect;
-      n(s.core.projectId);
+      const o = await this.connector.walletConnect;
+      n(o.core.projectId);
       return;
     }
     if (t.data.method === "walletConnect.disconnect") {
@@ -1055,8 +1059,8 @@ class L {
         if (!this.connector.walletConnect) throw new Error("WalletConnect is not configured");
         const a = await (await this.connector.walletConnect).disconnect(t.data.params);
         n(a);
-      } catch (s) {
-        o(s);
+      } catch (o) {
+        s(o);
       }
       return;
     }
@@ -1064,10 +1068,10 @@ class L {
       this.assertPermissions(e, "walletConnect", t);
       try {
         if (!this.connector.walletConnect) throw new Error("WalletConnect is not configured");
-        const s = await this.connector.walletConnect, a = s.session.keys[s.session.keys.length - 1], i = a ? s.session.get(a) : null;
+        const o = await this.connector.walletConnect, a = o.session.keys[o.session.keys.length - 1], i = a ? o.session.get(a) : null;
         n(i ? { topic: i.topic, namespaces: i.namespaces } : null);
-      } catch (s) {
-        o(s);
+      } catch (o) {
+        s(o);
       }
       return;
     }
@@ -1077,28 +1081,28 @@ class L {
         if (!this.connector.walletConnect) throw new Error("WalletConnect is not configured");
         const a = await (await this.connector.walletConnect).request(t.data.params);
         n(a);
-      } catch (s) {
-        o(s);
+      } catch (o) {
+        s(o);
       }
       return;
     }
     if (t.data.method === "external") {
       this.assertPermissions(e, "external", t);
       try {
-        const { entity: s, key: a, args: i } = t.data.params, l = s.split(".").reduce((w, v) => w[v], window);
-        s === "nightly.near" && a === "signTransaction" && (i[0].encode = () => i[0]);
+        const { entity: o, key: a, args: i } = t.data.params, l = o.split(".").reduce((w, v) => w[v], window);
+        o === "nightly.near" && a === "signTransaction" && (i[0].encode = () => i[0]);
         const c = typeof l[a] == "function" ? await l[a](...i || []) : l[a];
         n(c);
-      } catch (s) {
-        o(s);
+      } catch (o) {
+        s(o);
       }
       return;
     }
     if (t.data.method === "open") {
       this.assertPermissions(e, "allowsOpen", t);
-      const s = typeof window < "u" ? window?.Telegram?.WebApp : null;
-      if (s && t.data.params.url.startsWith("https://t.me")) {
-        s.openTelegramLink(t.data.params.url);
+      const o = typeof window < "u" ? window?.Telegram?.WebApp : null;
+      if (o && t.data.params.url.startsWith("https://t.me")) {
+        o.openTelegramLink(t.data.params.url);
         return;
       }
       const a = window.open(t.data.params.url, "_blank", t.data.params.features), i = a ? g() : null, l = (c) => {
@@ -1122,9 +1126,9 @@ class L {
     }
     if (t.data.method === "open.nativeApp") {
       this.assertPermissions(e, "allowsOpen", t);
-      const s = u(t.data.params.url);
-      if (!s || ["https", "http", "javascript:", "file:", "data:", "blob:", "about:"].includes(s.protocol))
-        throw o("Invalid URL"), new Error("[open.nativeApp] Invalid URL");
+      const o = u(t.data.params.url);
+      if (!o || ["https", "http", "javascript:", "file:", "data:", "blob:", "about:"].includes(o.protocol))
+        throw s("Invalid URL"), new Error("[open.nativeApp] Invalid URL");
       const i = document.createElement("iframe");
       i.src = t.data.params.url, i.style.display = "none", document.body.appendChild(i), e.postMessage({ ...t.data, status: "success", result: null });
       return;
@@ -1137,8 +1141,8 @@ class L {
     let n = u(e.manifest.executor);
     if (n || (n = u(location.origin + e.manifest.executor)), !n) throw new Error("Invalid executor URL");
     n.searchParams.set("nonce", N);
-    const o = await fetch(n.toString()).then((s) => s.text());
-    return this.connector.logger?.log("New version of code fetched"), this.actualCode = o, o === t ? (this.connector.logger?.log("New version of code is the same as the current version"), this.actualCode) : (await this.connector.db.setItem(`${this.manifest.id}:${this.manifest.version}`, o), this.connector.logger?.log("New version of code saved to cache"), o);
+    const s = await fetch(n.toString()).then((o) => o.text());
+    return this.connector.logger?.log("New version of code fetched"), this.actualCode = s, s === t ? (this.connector.logger?.log("New version of code is the same as the current version"), this.actualCode) : (await this.connector.db.setItem(`${this.manifest.id}:${this.manifest.version}`, s), this.connector.logger?.log("New version of code saved to cache"), s);
   }
   async loadCode() {
     const e = await this.connector.db.getItem(`${this.manifest.id}:${this.manifest.version}`).catch(() => null);
@@ -1150,15 +1154,15 @@ class L {
     this.connector.logger?.log("Add to queue", e, t), this.connector.logger?.log("Calling method", e, t);
     const n = await this.loadCode();
     this.connector.logger?.log("Code loaded, preparing");
-    const o = new M(this, n, this._onMessage);
-    this.connector.logger?.log("Code loaded, iframe initialized"), await o.readyPromise, this.connector.logger?.log("Iframe ready");
-    const s = g();
+    const s = new M(this, n, this._onMessage);
+    this.connector.logger?.log("Code loaded, iframe initialized"), await s.readyPromise, this.connector.logger?.log("Iframe ready");
+    const o = g();
     return new Promise((a, i) => {
       try {
         const l = (c) => {
-          c.data.id !== s || c.data.origin !== o.origin || (o.dispose(), window.removeEventListener("message", l), this.connector.logger?.log("postMessage", { result: c.data, request: { method: e, params: t } }), c.data.status === "failed" ? i(c.data.result) : a(c.data.result));
+          c.data.id !== o || c.data.origin !== s.origin || (s.dispose(), window.removeEventListener("message", l), this.connector.logger?.log("postMessage", { result: c.data, request: { method: e, params: t } }), c.data.status === "failed" ? i(c.data.result) : a(c.data.result));
         };
-        window.addEventListener("message", l), o.postMessage({ method: e, params: t, id: s }), o.on("close", () => i(new Error("Wallet closed")));
+        window.addEventListener("message", l), s.postMessage({ method: e, params: t, id: o }), s.on("close", () => i(new Error("Wallet closed")));
       } catch (l) {
         this.connector.logger?.log("Iframe error", l), i(l);
       }
@@ -1188,6 +1192,14 @@ class p {
       methodNames: e?.methodNames
     });
   }
+  async signInAndSignMessage(e) {
+    return this.executor.call("wallet:signInAndSignMessage", {
+      network: e?.network || this.connector.network,
+      contractId: e?.contractId,
+      methodNames: e?.methodNames,
+      messageParams: e.messageParams
+    });
+  }
   async signOut(e) {
     const t = { ...e, network: e?.network || this.connector.network };
     await this.executor.call("wallet:signOut", t), await this.executor.clearStorage();
@@ -1197,13 +1209,13 @@ class p {
     return this.executor.call("wallet:getAccounts", t);
   }
   async signAndSendTransaction(e) {
-    const t = d(e.actions), n = { ...e, actions: t, network: e.network || this.connector.network };
+    const t = h(e.actions), n = { ...e, actions: t, network: e.network || this.connector.network };
     return this.executor.call("wallet:signAndSendTransaction", n);
   }
   async signAndSendTransactions(e) {
-    const t = e.transactions.map((o) => ({
-      actions: d(o.actions),
-      receiverId: o.receiverId
+    const t = e.transactions.map((s) => ({
+      actions: h(s.actions),
+      receiverId: s.receiverId
     })), n = { ...e, transactions: t, network: e.network || this.connector.network };
     return this.executor.call("wallet:signAndSendTransactions", n);
   }
@@ -1216,7 +1228,7 @@ class p {
       ...e,
       delegateActions: e.delegateActions.map((n) => ({
         ...n,
-        actions: d(n.actions)
+        actions: h(n.actions)
       })),
       network: e.network || this.connector.network
     };
@@ -1237,6 +1249,14 @@ class W {
       methodNames: e?.methodNames
     });
   }
+  async signInAndSignMessage(e) {
+    return this.wallet.signInAndSignMessage({
+      network: e?.network || this.connector.network,
+      contractId: e?.contractId,
+      methodNames: e?.methodNames,
+      messageParams: e.messageParams
+    });
+  }
   async signOut(e) {
     await this.wallet.signOut({ network: e?.network || this.connector.network });
   }
@@ -1244,17 +1264,17 @@ class W {
     return this.wallet.getAccounts({ network: e?.network || this.connector.network });
   }
   async signAndSendTransaction(e) {
-    const t = d(e.actions), n = e.network || this.connector.network, o = await this.wallet.signAndSendTransaction({ ...e, actions: t, network: n });
-    if (!o) throw new Error("No result from wallet");
-    return Array.isArray(o.transactions) ? o.transactions[0] : o;
+    const t = h(e.actions), n = e.network || this.connector.network, s = await this.wallet.signAndSendTransaction({ ...e, actions: t, network: n });
+    if (!s) throw new Error("No result from wallet");
+    return Array.isArray(s.transactions) ? s.transactions[0] : s;
   }
   async signAndSendTransactions(e) {
-    const t = e.network || this.connector.network, n = e.transactions.map((s) => ({
-      actions: d(s.actions),
-      receiverId: s.receiverId
-    })), o = await this.wallet.signAndSendTransactions({ ...e, transactions: n, network: t });
-    if (!o) throw new Error("No result from wallet");
-    return Array.isArray(o.transactions) ? o.transactions : o;
+    const t = e.network || this.connector.network, n = e.transactions.map((o) => ({
+      actions: h(o.actions),
+      receiverId: o.receiverId
+    })), s = await this.wallet.signAndSendTransactions({ ...e, transactions: n, network: t });
+    if (!s) throw new Error("No result from wallet");
+    return Array.isArray(s.transactions) ? s.transactions : s;
   }
   async signMessage(e) {
     return this.wallet.signMessage({ ...e, network: e.network || this.connector.network });
@@ -1264,7 +1284,7 @@ class W {
       ...e,
       delegateActions: e.delegateActions.map((t) => ({
         ...t,
-        actions: d(t.actions)
+        actions: h(t.actions)
       })),
       network: e.network || this.connector.network
     });
@@ -1283,16 +1303,17 @@ const D = {
   features: {
     signMessage: !0,
     signInWithoutAddKey: !0,
+    signInAndSignMessage: !0,
     signAndSendTransaction: !0,
     signAndSendTransactions: !0,
-    signDelegateAction: !0
+    signDelegateActions: !0
   },
   permissions: {
     storage: !0,
     allowsOpen: []
   }
 };
-class T extends b {
+class j extends b {
   constructor(e) {
     super(e), this.delegate = e, this.update({ wallets: e.wallets, showSettings: !1 });
   }
@@ -1302,7 +1323,7 @@ class T extends b {
     }), this.root.querySelectorAll(".remove-wallet-button").forEach((e) => {
       e instanceof SVGSVGElement && this.addListener(e, "click", async (t) => {
         t.stopPropagation(), await this.delegate.onRemoveDebugManifest(e.dataset.type);
-        const n = this.state.wallets.filter((o) => o.id !== e.dataset.type);
+        const n = this.state.wallets.filter((s) => s.id !== e.dataset.type);
         this.update({ wallets: n });
       });
     }), this.addListener(".add-debug-manifest-button", "click", async () => {
@@ -1318,7 +1339,7 @@ class T extends b {
     super.create({ show: !0 });
   }
   walletDom(e) {
-    const t = h`
+    const t = d`
       <svg
         class="remove-wallet-button"
         data-type="${e.id}"
@@ -1333,7 +1354,7 @@ class T extends b {
         <path d="M6 6L18 18" stroke="rgba(255,255,255,0.5)" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" />
       </svg>
     `;
-    return h`
+    return d`
       <div class="connect-item" data-type="${e.id}">
         <img style="background: #333" src="${e.icon}" alt="${e.name}" />
         <div class="connect-item-info">
@@ -1346,17 +1367,17 @@ class T extends b {
   }
   get footer() {
     if (!this.delegate.footer) return "";
-    const { icon: e, heading: t, link: n, linkText: o } = this.delegate.footer;
-    return h`
+    const { icon: e, heading: t, link: n, linkText: s } = this.delegate.footer;
+    return d`
       <div class="footer">
-        <img src="${e}" alt="${t}" />
+        ${e ? d`<img src="${e}" alt="${t}" />` : ""}
         <p>${t}</p>
-        <a class="get-wallet-link" href="${n}" target="_blank">${o}</a>
+        <a class="get-wallet-link" href="${n}" target="_blank">${s}</a>
       </div>
     `;
   }
   get dom() {
-    return this.state.showSettings ? h`
+    return this.state.showSettings ? d`
         <div class="modal-container">
           <div class="modal-content">
             <div class="modal-header">
@@ -1381,7 +1402,7 @@ class T extends b {
             ${this.footer}
           </div>
         </div>
-      ` : h`<div class="modal-container">
+      ` : d`<div class="modal-container">
       <div class="modal-content">
         <div class="modal-header">
           <p>Select wallet</p>
@@ -1401,7 +1422,7 @@ class T extends b {
     </div>`;
   }
 }
-class j {
+class T {
   dbName;
   storeName;
   version;
@@ -1415,13 +1436,13 @@ class j {
         return;
       }
       const n = indexedDB.open(this.dbName, this.version);
-      n.onerror = (o) => {
-        console.error("Error opening database:", o.target.error), t(new Error("Error opening database"));
-      }, n.onsuccess = (o) => {
+      n.onerror = (s) => {
+        console.error("Error opening database:", s.target.error), t(new Error("Error opening database"));
+      }, n.onsuccess = (s) => {
         e(n.result);
-      }, n.onupgradeneeded = (o) => {
-        const s = n.result;
-        s.objectStoreNames.contains(this.storeName) || s.createObjectStore(this.storeName);
+      }, n.onupgradeneeded = (s) => {
+        const o = n.result;
+        o.objectStoreNames.contains(this.storeName) || o.createObjectStore(this.storeName);
       };
     });
   }
@@ -1429,15 +1450,15 @@ class j {
     const t = await this.getDb();
     if (typeof e == "number" && (e = e.toString()), typeof e != "string")
       throw new Error("Key must be a string");
-    return new Promise((n, o) => {
+    return new Promise((n, s) => {
       if (!this.storeName) {
-        o(new Error("Store name not set"));
+        s(new Error("Store name not set"));
         return;
       }
-      const s = t.transaction(this.storeName, "readonly");
-      s.onerror = (l) => o(s.error);
-      const i = s.objectStore(this.storeName).get(e);
-      i.onerror = (l) => o(i.error), i.onsuccess = () => {
+      const o = t.transaction(this.storeName, "readonly");
+      o.onerror = (l) => s(o.error);
+      const i = o.objectStore(this.storeName).get(e);
+      i.onerror = (l) => s(i.error), i.onsuccess = () => {
         n(i.result), t.close();
       };
     });
@@ -1446,16 +1467,16 @@ class j {
     const n = await this.getDb();
     if (typeof e == "number" && (e = e.toString()), typeof e != "string")
       throw new Error("Key must be a string");
-    return new Promise((o, s) => {
+    return new Promise((s, o) => {
       if (!this.storeName) {
-        s(new Error("Store name not set"));
+        o(new Error("Store name not set"));
         return;
       }
       const a = n.transaction(this.storeName, "readwrite");
-      a.onerror = (c) => s(a.error);
+      a.onerror = (c) => o(a.error);
       const l = a.objectStore(this.storeName).put(t, e);
-      l.onerror = (c) => s(l.error), l.onsuccess = () => {
-        n.close(), o();
+      l.onerror = (c) => o(l.error), l.onsuccess = () => {
+        n.close(), s();
       };
     });
   }
@@ -1463,15 +1484,15 @@ class j {
     const t = await this.getDb();
     if (typeof e == "number" && (e = e.toString()), typeof e != "string")
       throw new Error("Key must be a string");
-    return new Promise((n, o) => {
+    return new Promise((n, s) => {
       if (!this.storeName) {
-        o(new Error("Store name not set"));
+        s(new Error("Store name not set"));
         return;
       }
-      const s = t.transaction(this.storeName, "readwrite");
-      s.onerror = (l) => o(s.error);
-      const i = s.objectStore(this.storeName).delete(e);
-      i.onerror = (l) => o(i.error), i.onsuccess = () => {
+      const o = t.transaction(this.storeName, "readwrite");
+      o.onerror = (l) => s(o.error);
+      const i = o.objectStore(this.storeName).delete(e);
+      i.onerror = (l) => s(i.error), i.onsuccess = () => {
         t.close(), n();
       };
     });
@@ -1483,9 +1504,9 @@ class j {
         n(new Error("Store name not set"));
         return;
       }
-      const o = e.transaction(this.storeName, "readonly");
-      o.onerror = (i) => n(o.error);
-      const a = o.objectStore(this.storeName).getAllKeys();
+      const s = e.transaction(this.storeName, "readonly");
+      s.onerror = (i) => n(s.error);
+      const a = s.objectStore(this.storeName).getAllKeys();
       a.onerror = (i) => n(a.error), a.onsuccess = () => {
         t(a.result), e.close();
       };
@@ -1498,9 +1519,9 @@ class j {
         n(new Error("Store name not set"));
         return;
       }
-      const o = e.transaction(this.storeName, "readonly");
-      o.onerror = (i) => n(o.error);
-      const a = o.objectStore(this.storeName).count();
+      const s = e.transaction(this.storeName, "readonly");
+      s.onerror = (i) => n(s.error);
+      const a = s.objectStore(this.storeName).count();
       a.onerror = (i) => n(a.error), a.onsuccess = () => {
         t(a.result), e.close();
       };
@@ -1516,9 +1537,9 @@ class j {
         n(new Error("Store name not set"));
         return;
       }
-      const o = e.transaction(this.storeName, "readwrite");
-      o.onerror = (i) => n(o.error);
-      const a = o.objectStore(this.storeName).clear();
+      const s = e.transaction(this.storeName, "readwrite");
+      s.onerror = (i) => n(s.error);
+      const a = s.objectStore(this.storeName).clear();
       a.onerror = (i) => n(a.error), a.onsuccess = () => {
         e.close(), t();
       };
@@ -1529,6 +1550,9 @@ const O = [
   "https://raw.githubusercontent.com/hot-dao/near-selector/refs/heads/main/repository/manifest.json",
   "https://cdn.jsdelivr.net/gh/azbang/hot-connector/repository/manifest.json"
 ];
+function K(r) {
+  return (e) => Object.entries(r).every(([t, n]) => !(n && !e.manifest.features?.[t]));
+}
 class q {
   storage;
   events;
@@ -1546,26 +1570,26 @@ class q {
   autoConnect;
   whenManifestLoaded;
   constructor(e) {
-    this.db = new j("hot-connector", "wallets"), this.storage = e?.storage ?? new k(), this.events = e?.events ?? new y(), this.logger = e?.logger, this.network = e?.network ?? "mainnet", this.walletConnect = e?.walletConnect, this.autoConnect = e?.autoConnect ?? !0, this.providers = e?.providers ?? { mainnet: [], testnet: [] }, this.excludedWallets = e?.excludedWallets ?? [], this.features = e?.features ?? {}, this.signInData = e?.signIn, e?.footerBranding !== void 0 ? this.footerBranding = e?.footerBranding : this.footerBranding = {
-      icon: "https://tgapp.herewallet.app/images/hot/hot-icon.png",
+    this.db = new T("hot-connector", "wallets"), this.storage = e?.storage ?? new k(), this.events = e?.events ?? new y(), this.logger = e?.logger, this.network = e?.network ?? "mainnet", this.walletConnect = e?.walletConnect, this.autoConnect = e?.autoConnect ?? !0, this.providers = e?.providers ?? { mainnet: [], testnet: [] }, this.excludedWallets = e?.excludedWallets ?? [], this.features = e?.features ?? {}, this.signInData = e?.signIn, e?.footerBranding !== void 0 ? this.footerBranding = e?.footerBranding : this.footerBranding = {
+      icon: "https://pages.near.org/wp-content/uploads/2023/11/NEAR_token.png",
       heading: "NEAR Connector",
       link: "https://wallet.near.org",
       linkText: "Don't have a wallet?"
     }, this.whenManifestLoaded = new Promise(async (t) => {
       e?.manifest == null || typeof e.manifest == "string" ? this.manifest = await this._loadManifest(e?.manifest).catch(() => ({ wallets: [], version: "1.0.0" })) : this.manifest = e?.manifest ?? { wallets: [], version: "1.0.0" };
       const n = new Set(this.excludedWallets);
-      n.delete("hot-wallet"), this.manifest.wallets = this.manifest.wallets.filter((o) => !(o.permissions.walletConnect && !this.walletConnect || n.has(o.id))), await new Promise((o) => setTimeout(o, 100)), t();
+      n.delete("hot-wallet"), this.manifest.wallets = this.manifest.wallets.filter((s) => !(s.permissions.walletConnect && !this.walletConnect || n.has(s.id))), await new Promise((s) => setTimeout(s, 100)), t();
     }), typeof window < "u" && (window.addEventListener("near-wallet-injected", this._handleNearWalletInjected), window.dispatchEvent(new Event("near-selector-ready")), window.addEventListener("message", async (t) => {
       t.data.type === "near-wallet-injected" && (await this.whenManifestLoaded.catch(() => {
-      }), this.wallets = this.wallets.filter((n) => n.manifest.id !== t.data.manifest.id), this.wallets.unshift(new I(this, t.data.manifest)), this.events.emit("selector:walletsChanged", {}), this.autoConnect && this.connect(t.data.manifest.id));
+      }), this.wallets = this.wallets.filter((n) => n.manifest.id !== t.data.manifest.id), this.wallets.unshift(new A(this, t.data.manifest)), this.events.emit("selector:walletsChanged", {}), this.autoConnect && this.connect({ walletId: t.data.manifest.id }));
     })), this.whenManifestLoaded.then(() => {
       typeof window < "u" && window.parent.postMessage({ type: "near-selector-ready" }, "*"), this.manifest.wallets.forEach((t) => this.registerWallet(t)), this.storage.get("debug-wallets").then((t) => {
-        JSON.parse(t ?? "[]").forEach((o) => this.registerDebugWallet(o));
+        JSON.parse(t ?? "[]").forEach((s) => this.registerDebugWallet(s));
       });
     });
   }
   get availableWallets() {
-    return this.wallets.filter((t) => Object.entries(this.features).every(([n, o]) => !(o && !t.manifest.features?.[n]))).filter((t) => !(this.network === "testnet" && !t.manifest.features?.testnet));
+    return this.wallets.filter((t) => Object.entries(this.features).every(([n, s]) => !(s && !t.manifest.features?.[n]))).filter((t) => !(this.network === "testnet" && !t.manifest.features?.testnet));
   }
   _handleNearWalletInjected = (e) => {
     this.wallets = this.wallets.filter((t) => t.manifest.id !== e.detail.manifest.id), this.wallets.unshift(new W(this, e.detail)), this.events.emit("selector:walletsChanged", {});
@@ -1573,9 +1597,9 @@ class q {
   async _loadManifest(e) {
     const t = e ? [e] : O;
     for (const n of t) {
-      const o = await fetch(n).catch(() => null);
-      if (!(!o || !o.ok))
-        return await o.json();
+      const s = await fetch(n).catch(() => null);
+      if (!(!s || !s.ok))
+        return await s.json();
     }
     throw new Error("Failed to load manifest");
   }
@@ -1598,9 +1622,9 @@ class q {
     if (!t.executor) throw new Error("Manifest must have an executor");
     if (!t.features) throw new Error("Manifest must have features");
     if (!t.permissions) throw new Error("Manifest must have permissions");
-    if (this.wallets.find((o) => o.manifest.id === t.id)) throw new Error("Wallet already registered");
+    if (this.wallets.find((s) => s.manifest.id === t.id)) throw new Error("Wallet already registered");
     t.debug = !0, this.wallets.unshift(new p(this, t)), this.events.emit("selector:walletsChanged", {});
-    const n = this.wallets.filter((o) => o.manifest.debug).map((o) => o.manifest);
+    const n = this.wallets.filter((s) => s.manifest.debug).map((s) => s.manifest);
     return this.storage.set("debug-wallets", JSON.stringify(n)), t;
   }
   async removeDebugWallet(e) {
@@ -1608,35 +1632,56 @@ class q {
     const t = this.wallets.filter((n) => n.manifest.debug).map((n) => n.manifest);
     this.storage.set("debug-wallets", JSON.stringify(t)), this.events.emit("selector:walletsChanged", {});
   }
-  async selectWallet() {
+  async selectWallet({ features: e = {} } = {}) {
     return await this.whenManifestLoaded.catch(() => {
-    }), new Promise((e, t) => {
-      const n = new T({
+    }), new Promise((t, n) => {
+      const s = new j({
         footer: this.footerBranding,
-        wallets: this.availableWallets.map((o) => o.manifest),
+        wallets: this.availableWallets.filter(K(e)).map((o) => o.manifest),
         onRemoveDebugManifest: async (o) => this.removeDebugWallet(o),
         onAddDebugManifest: async (o) => this.registerDebugWallet(o),
-        onReject: () => (t(new Error("User rejected")), n.destroy()),
-        onSelect: (o) => (e(o), n.destroy())
+        onReject: () => (n(new Error("User rejected")), s.destroy()),
+        onSelect: (o) => (t(o), s.destroy())
       });
-      n.create();
+      s.create();
     });
   }
-  async connect(e) {
+  async connect(e = {}) {
+    let t = e.walletId;
+    const n = e.signMessageParams;
     await this.whenManifestLoaded.catch(() => {
-    }), e || (e = await this.selectWallet());
+    }), t || (t = await this.selectWallet(e.signMessageParams != null ? { features: { signInAndSignMessage: !0 } } : void 0));
     try {
-      const t = await this.wallet(e);
-      this.logger?.log("Wallet available to connect", t), await this.storage.set("selected-wallet", e), this.logger?.log("Set preferred wallet, try to signIn", e);
-      const n = await t.signIn({
-        contractId: this.signInData?.contractId,
-        methodNames: this.signInData?.methodNames,
-        network: this.network
-      });
-      if (!n?.length) throw new Error("Failed to sign in");
-      return this.logger?.log("Signed in to wallet", e, n), this.events.emit("wallet:signIn", { wallet: t, accounts: n, success: !0 }), t;
-    } catch (t) {
-      throw this.logger?.log("Failed to connect to wallet", t), t;
+      const s = await this.wallet(t);
+      if (this.logger?.log("Wallet available to connect", s), await this.storage.set("selected-wallet", t), this.logger?.log(`Set preferred wallet, try to signIn${n != null ? " (with signed message)" : ""}`, t), n != null) {
+        const o = await s.signInAndSignMessage({
+          contractId: this.signInData?.contractId,
+          methodNames: this.signInData?.methodNames,
+          messageParams: n,
+          network: this.network
+        });
+        if (!o?.length) throw new Error("Failed to sign in");
+        this.logger?.log("Signed in to wallet (with signed message)", t, o), this.events.emit("wallet:signInAndSignMessage", { wallet: s, accounts: o, success: !0 }), this.events.emit("wallet:signIn", {
+          wallet: s,
+          accounts: o.map((a) => ({
+            accountId: a.accountId,
+            publicKey: a.publicKey
+          })),
+          success: !0,
+          source: "signInAndSignMessage"
+        });
+      } else {
+        const o = await s.signIn({
+          contractId: this.signInData?.contractId,
+          methodNames: this.signInData?.methodNames,
+          network: this.network
+        });
+        if (!o?.length) throw new Error("Failed to sign in");
+        this.logger?.log("Signed in to wallet", t, o), this.events.emit("wallet:signIn", { wallet: s, accounts: o, success: !0, source: "signIn" });
+      }
+      return s;
+    } catch (s) {
+      throw this.logger?.log("Failed to connect to wallet", s), s;
     }
   }
   async disconnect(e) {
@@ -1645,7 +1690,7 @@ class q {
   async getConnectedWallet() {
     await this.whenManifestLoaded.catch(() => {
     });
-    const e = await this.storage.get("selected-wallet"), t = this.wallets.find((o) => o.manifest.id === e);
+    const e = await this.storage.get("selected-wallet"), t = this.wallets.find((s) => s.manifest.id === e);
     if (!t) throw new Error("No wallet selected");
     const n = await t.getAccounts();
     if (!n?.length) throw new Error("No accounts found");
@@ -1664,10 +1709,10 @@ class q {
   async use(e) {
     await this.whenManifestLoaded.catch(() => {
     }), this.wallets = this.wallets.map((t) => new Proxy(t, {
-      get(n, o, s) {
-        const a = Reflect.get(n, o, s);
-        if (o in e && typeof a == "function") {
-          const i = e[o];
+      get(n, s, o) {
+        const a = Reflect.get(n, s, o);
+        if (s in e && typeof a == "function") {
+          const i = e[s];
           return function(...l) {
             const c = () => a.apply(n, l);
             return l.length > 0 ? i.call(this, ...l, c) : i.call(this, void 0, c);
@@ -1694,7 +1739,7 @@ export {
   W as InjectedWallet,
   k as LocalStorage,
   q as NearConnector,
-  I as ParentFrameWallet,
+  A as ParentFrameWallet,
   p as SandboxWallet,
-  d as nearActionsToConnectorActions
+  h as nearActionsToConnectorActions
 };