+++ src/popups/IframeWalletPopup.ts
@@ -30,26 +30,21 @@ export class IframeWalletPopup extends Popup<{}> {
 
   get footer() {
     if (!this.delegate.footer) return "";
-    const { icon, heading, link, linkText } = this.delegate.footer;
+    const { icon, heading } = this.delegate.footer;
 
     return html`
       <div class="footer">
-        <img src="${icon}" alt="${heading}" />
+        ${icon ? html`<img src="${icon}" alt="${heading}" />` : ""}
         <p>${heading}</p>
-        <a class="get-wallet-link" href="${link}" target="_blank">${linkText}</a>
       </div>
     `;
   }
 
   get dom() {
-    return html` <div class="modal-container">
+    return html`<div class="modal-container">
       <div class="modal-content">
         <div class="modal-body" style="padding: 0; overflow: auto;"></div>
-        <div class="footer">
-          <img src="https://tgapp.herewallet.app/images/hot/hot-icon.png" alt="NEAR Connector" />
-          <p>NEAR Connector</p>
-          <p class="get-wallet-link">Don't have a wallet?</p>
-        </div>
+        ${this.footer}
       </div>
     </div>`;
   }